requirements.txt
PyHamcrest==1.9.0
pytest==4.2.0
requests==2.21.0
